+++ integrationstests/src/main/java/de/raysha/lib/jsimpleshell/DependencyInjectionCommand.java
@@ -9,15 +9,15 @@ import de.raysha.lib.jsimpleshell.io.OutputBuilder;
 public class DependencyInjectionCommand {
 
 	@Inject
-	private Shell shell;
+	protected Shell shell;
 
-	private InputBuilder input;
+	protected InputBuilder input;
 
 	@Inject
-	private OutputBuilder output;
+	protected OutputBuilder output;
 
 	@javax.inject.Inject
-	private MessageResolver messageResolver;
+	protected MessageResolver messageResolver;
 
 	@Command
 	public String checkDependencies(){
@@ -49,5 +49,28 @@ public class DependencyInjectionCommand {
 	public void setIO(InputBuilder input, OutputBuilder output) {
 		this.input = input;
 		this.output = output;
+	}
+
+	public static class InheritedDependencyInjectionCommand extends DependencyInjectionCommand {
+		@Command
+		public String checkInheritedDependencies(){
+			if(shell == null){
+				return "shell is null!";
+			}
+
+			if(input == null){
+				return "input is null!";
+			}
+
+			if(output == null){
+				return "output is null!";
+			}
+
+			if(messageResolver == null){
+				return "messageResolver is null!";
+			}
+
+			return null;
+		}
 	}
 }